requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==3.2
django-mysql==4.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
inflection==0.5.1
install==1.3.5
jsonschema==4.14.0
mysqlclient==2.1.1
PyJWT==2.4.0
pyrsistent==0.18.1
pytz==2022.2.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
